fix: Fixing ProfileHeader wip
