--- src/components/shared/ImageModal.tsx
@@ -1,74 +1,0 @@
-import React, { useRef, useState } from "react";
-
-interface ImageUploadModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-  onUpload: (file: File) => void;
-}
-
-const ImageUploadModal: React.FC<ImageUploadModalProps> = ({
-  isOpen,
-  onClose,
-  onUpload,
-}) => {
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const [selectedFile, setSelectedFile] = useState<File | null>(null);
-
-  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files && e.target.files[0];
-    setSelectedFile(file);
-  };
-
-  const handleUpload = () => {
-    if (selectedFile) {
-      onUpload(selectedFile);
-      setSelectedFile(null);
-      onClose();
-    }
-  };
-
-  const handleClose = () => {
-    setSelectedFile(null);
-    onClose();
-  };
-
-  return (
-    <>
-      {isOpen && (
-        <div className="fixed top-0 left-0 w-full h-full flex justify-center items-center bg-black bg-opacity-50">
-          <div className="bg-white p-6 rounded-lg shadow-lg">
-            <h2 className="text-xl font-bold mb-4">Nová fotka</h2>
-            <input
-              type="file"
-              accept="image/*"
-              onChange={handleFileChange}
-              ref={fileInputRef}
-            />
-            <div className="flex justify-between mt-4">
-              <button
-                className="px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-600"
-                onClick={() => fileInputRef.current?.click()}
-              >
-                Vybrať obrázok
-              </button>
-              <button
-                className="px-4 py-2 bg-red-500 text-white rounded hover:bg-red-600"
-                onClick={handleClose}
-              >
-                Zavrieť
-              </button>
-            </div>
-            <button
-              className="mt-4 px-4 py-2 bg-green-500 text-white rounded hover:bg-green-600"
-              onClick={handleUpload}
-            >
-              Nahrať
-            </button>
-          </div>
-        </div>
-      )}
-    </>
-  );
-};
-
-export default ImageUploadModal;
