+++ src/components/student/profile/ProfileHeader.tsx
@@ -1,8 +1,7 @@
 import Image from "next/image";
-import { useState } from "react";
+import { useState, useRef } from "react";
 import AvatarImage from "../../../images/noImage.png";
 import useStudent from "@/hooks/useStudent";
-import ImageUploadModal from "@/components/shared/ImageModal";
 import axios from "axios";
 
 const backendURL = process.env.NEXT_PUBLIC_BACKEND_URL as unknown as string;
@@ -12,10 +11,17 @@ const ProfileHeader: React.FC = () => {
   const [isLoading, setIsLoading] = useState(false);
   const [isModalOpen, setIsModalOpen] = useState(false);
   const { studentPersonalInfo } = useStudent();
+  const fileInputRef = useRef<HTMLInputElement>(null);
 
   const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const file = e.target.files && e.target.files[0];
     setSelectedFile(file);
+  };
+
+  
+  const handleClose = () => {
+    setSelectedFile(null);
+    setIsModalOpen(false); 
   };
 
   const handleUpload = async () => {
@@ -73,11 +79,37 @@ const ProfileHeader: React.FC = () => {
         >
           Nová fotka
         </button>
-        <ImageUploadModal
-          isOpen={isModalOpen}
-          onClose={() => setIsModalOpen(false)}
-          onUpload={handleUpload}
-        />
+        <div className="fixed top-0 left-0 w-full h-full flex justify-center items-center bg-black bg-opacity-50">
+          <div className="bg-white p-6 rounded-lg shadow-lg">
+            <h2 className="text-xl font-bold mb-4">Nová fotka</h2>
+            <input
+              type="file"
+              accept="image/*"
+              onChange={handleFileChange}
+              ref={fileInputRef}
+            />
+            <div className="flex justify-between mt-4">
+              <button
+                className="px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-600"
+                onClick={() => fileInputRef.current?.click()}
+              >
+                Vybrať obrázok
+              </button>
+              <button
+                className="px-4 py-2 bg-red-500 text-white rounded hover:bg-red-600"
+                onClick={handleClose}
+              >
+                Zavrieť
+              </button>
+            </div>
+            <button
+              className="mt-4 px-4 py-2 bg-green-500 text-white rounded hover:bg-green-600"
+              onClick={handleUpload}
+            >
+              Nahrať
+            </button>
+          </div>
+        </div>
       </div>
     </div>
   );
